# backend/openapi.json
{"openapi":"3.1.0","info":{"title":"HRMS Lite","description":"Production-grade HRMS Lite system for employee management, attendance tracking, filtering, and summary dashboard.","version":"1.0.0"},"paths":{"/api/v1/employees":{"post":{"tags":["Employees"],"summary":"Create a new employee","operationId":"create_employee_api_v1_employees_post","requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/EmployeeCreate"}}}},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/EmployeeResponse"}}}},"409":{"description":"Email or employee_code conflict"},"422":{"description":"Validation error"}}},"get":{"tags":["Employees"],"summary":"List employees with pagination and filters","operationId":"list_employees_api_v1_employees_get","parameters":[{"name":"page","in":"query","required":false,"schema":{"type":"integer","minimum":1,"default":1,"title":"Page"}},{"name":"per_page","in":"query","required":false,"schema":{"type":"integer","maximum":100,"minimum":1,"default":20,"title":"Per Page"}},{"name":"department","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Department"}},{"name":"is_active","in":"query","required":false,"schema":{"anyOf":[{"type":"boolean"},{"type":"null"}],"title":"Is Active"}},{"name":"search","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"Search name, email, or code","title":"Search"},"description":"Search name, email, or code"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PaginatedResponse_EmployeeResponse_"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/employees/{employee_id}":{"get":{"tags":["Employees"],"summary":"Get employee by ID","operationId":"get_employee_api_v1_employees__employee_id__get","parameters":[{"name":"employee_id","in":"path","required":true,"schema":{"type":"string","title":"Employee Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/EmployeeResponse"}}}},"404":{"description":"Employee not found"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"put":{"tags":["Employees"],"summary":"Update employee","operationId":"update_employee_api_v1_employees__employee_id__put","parameters":[{"name":"employee_id","in":"path","required":true,"schema":{"type":"string","title":"Employee Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/EmployeeUpdate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/EmployeeResponse"}}}},"404":{"description":"Not found"},"409":{"description":"Email conflict"},"422":{"description":"Validation error"}}},"delete":{"tags":["Employees"],"summary":"Delete employee (cascades attendance)","operationId":"delete_employee_api_v1_employees__employee_id__delete","parameters":[{"name":"employee_id","in":"path","required":true,"schema":{"type":"string","title":"Employee Id"}}],"responses":{"204":{"description":"Successful Response"},"404":{"description":"Employee not found"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/attendance":{"post":{"tags":["Attendance"],"summary":"Mark attendance for an employee","operationId":"mark_attendance_api_v1_attendance_post","requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AttendanceCreate"}}}},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AttendanceResponse"}}}},"404":{"description":"Employee not found"},"409":{"description":"Attendance already exists for this date"},"422":{"description":"Validation: future date, date before joining, invalid status"}}},"get":{"tags":["Attendance"],"summary":"List attendance records with filters","operationId":"list_attendance_api_v1_attendance_get","parameters":[{"name":"page","in":"query","required":false,"schema":{"type":"integer","minimum":1,"default":1,"title":"Page"}},{"name":"per_page","in":"query","required":false,"schema":{"type":"integer","maximum":100,"minimum":1,"default":20,"title":"Per Page"}},{"name":"employee_id","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Employee Id"}},{"name":"date","in":"query","required":false,"schema":{"anyOf":[{"type":"string","format":"date"},{"type":"null"}],"title":"Date"}},{"name":"date_from","in":"query","required":false,"schema":{"anyOf":[{"type":"string","format":"date"},{"type":"null"}],"title":"Date From"}},{"name":"date_to","in":"query","required":false,"schema":{"anyOf":[{"type":"string","format":"date"},{"type":"null"}],"title":"Date To"}},{"name":"status","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Status"}},{"name":"department","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Department"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PaginatedResponse_AttendanceResponse_"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/attendance/{attendance_id}":{"get":{"tags":["Attendance"],"summary":"Get attendance record by ID","operationId":"get_attendance_api_v1_attendance__attendance_id__get","parameters":[{"name":"attendance_id","in":"path","required":true,"schema":{"type":"string","title":"Attendance Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AttendanceResponse"}}}},"404":{"description":"Record not found"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"put":{"tags":["Attendance"],"summary":"Update attendance record (employee_id and date are immutable)","operationId":"update_attendance_api_v1_attendance__attendance_id__put","parameters":[{"name":"attendance_id","in":"path","required":true,"schema":{"type":"string","title":"Attendance Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AttendanceUpdate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AttendanceResponse"}}}},"404":{"description":"Record not found"},"422":{"description":"Validation error"}}},"delete":{"tags":["Attendance"],"summary":"Delete attendance record","operationId":"delete_attendance_api_v1_attendance__attendance_id__delete","parameters":[{"name":"attendance_id","in":"path","required":true,"schema":{"type":"string","title":"Attendance Id"}}],"responses":{"204":{"description":"Successful Response"},"404":{"description":"Record not found"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/dashboard/summary":{"get":{"tags":["Dashboard"],"summary":"Get aggregated attendance summary","description":"Returns attendance counts, rates, and department breakdown. Excludes inactive employees by default (INV-11). Set include_inactive=true to include them.","operationId":"get_summary_api_v1_dashboard_summary_get","parameters":[{"name":"date_from","in":"query","required":false,"schema":{"anyOf":[{"type":"string","format":"date"},{"type":"null"}],"description":"Start date (defaults to today)","title":"Date From"},"description":"Start date (defaults to today)"},{"name":"date_to","in":"query","required":false,"schema":{"anyOf":[{"type":"string","format":"date"},{"type":"null"}],"description":"End date (defaults to date_from)","title":"Date To"},"description":"End date (defaults to date_from)"},{"name":"department","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Department"}},{"name":"include_inactive","in":"query","required":false,"schema":{"type":"boolean","description":"Include inactive employees","default":false,"title":"Include Inactive"},"description":"Include inactive employees"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DashboardSummaryResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/health":{"get":{"tags":["Health"],"summary":"Health Check","operationId":"health_check_api_v1_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"AttendanceCreate":{"properties":{"employee_id":{"type":"string","title":"Employee Id","description":"UUID of the employee"},"date":{"type":"string","format":"date","title":"Date"},"status":{"$ref":"#/components/schemas/AttendanceStatus"},"check_in":{"anyOf":[{"type":"string","format":"time"},{"type":"null"}],"title":"Check In"},"check_out":{"anyOf":[{"type":"string","format":"time"},{"type":"null"}],"title":"Check Out"},"notes":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Notes"}},"type":"object","required":["employee_id","date","status"],"title":"AttendanceCreate","description":"Request schema for marking attendance."},"AttendanceResponse":{"properties":{"id":{"type":"string","title":"Id"},"employee_id":{"type":"string","title":"Employee Id"},"employee_name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Employee Name"},"employee_code":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Employee Code"},"date":{"type":"string","format":"date","title":"Date"},"status":{"type":"string","title":"Status"},"check_in":{"anyOf":[{"type":"string","format":"time"},{"type":"null"}],"title":"Check In"},"check_out":{"anyOf":[{"type":"string","format":"time"},{"type":"null"}],"title":"Check Out"},"notes":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Notes"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"updated_at":{"type":"string","format":"date-time","title":"Updated At"}},"type":"object","required":["id","employee_id","date","status","check_in","check_out","notes","created_at","updated_at"],"title":"AttendanceResponse","description":"Response schema for attendance data, includes denormalized employee info."},"AttendanceStatus":{"type":"string","enum":["PRESENT","ABSENT","HALF_DAY","ON_LEAVE"],"title":"AttendanceStatus","description":"Closed set of attendance status values (INV-6)."},"AttendanceUpdate":{"properties":{"status":{"anyOf":[{"$ref":"#/components/schemas/AttendanceStatus"},{"type":"null"}]},"check_in":{"anyOf":[{"type":"string","format":"time"},{"type":"null"}],"title":"Check In"},"check_out":{"anyOf":[{"type":"string","format":"time"},{"type":"null"}],"title":"Check Out"},"notes":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Notes"}},"type":"object","title":"AttendanceUpdate","description":"Request schema for updating attendance. employee_id and date are immutable."},"DashboardSummaryResponse":{"properties":{"date_range":{"$ref":"#/components/schemas/DateRange"},"total_employees":{"type":"integer","title":"Total Employees"},"summary":{"$ref":"#/components/schemas/StatusSummary"},"attendance_rate":{"type":"number","title":"Attendance Rate"},"department_breakdown":{"items":{"$ref":"#/components/schemas/DepartmentBreakdown"},"type":"array","title":"Department Breakdown"}},"type":"object","required":["date_range","total_employees","summary","attendance_rate","department_breakdown"],"title":"DashboardSummaryResponse","description":"Aggregated dashboard summary (Section 5.2.3 of design)."},"DateRange":{"properties":{"date_from":{"type":"string","format":"date","title":"Date From"},"date_to":{"type":"string","format":"date","title":"Date To"}},"type":"object","required":["date_from","date_to"],"title":"DateRange","description":"Date range for the dashboard query."},"DepartmentBreakdown":{"properties":{"department":{"type":"string","title":"Department"},"present":{"type":"integer","title":"Present","default":0},"absent":{"type":"integer","title":"Absent","default":0},"half_day":{"type":"integer","title":"Half Day","default":0},"on_leave":{"type":"integer","title":"On Leave","default":0}},"type":"object","required":["department"],"title":"DepartmentBreakdown","description":"Per-department attendance breakdown."},"EmployeeCreate":{"properties":{"employee_code":{"type":"string","maxLength":20,"minLength":1,"title":"Employee Code","description":"Unique business identifier (e.g., EMP-001)"},"name":{"type":"string","maxLength":100,"minLength":1,"title":"Name"},"email":{"type":"string","maxLength":255,"minLength":5,"title":"Email","description":"Unique email address"},"department":{"type":"string","maxLength":100,"minLength":1,"title":"Department"},"designation":{"anyOf":[{"type":"string","maxLength":100},{"type":"null"}],"title":"Designation"},"date_of_joining":{"type":"string","format":"date","title":"Date Of Joining"},"phone":{"anyOf":[{"type":"string","maxLength":20},{"type":"null"}],"title":"Phone"}},"type":"object","required":["employee_code","name","email","department","date_of_joining"],"title":"EmployeeCreate","description":"Request schema for creating an employee."},"EmployeeResponse":{"properties":{"id":{"type":"string","title":"Id"},"employee_code":{"type":"string","title":"Employee Code"},"name":{"type":"string","title":"Name"},"email":{"type":"string","title":"Email"},"department":{"type":"string","title":"Department"},"designation":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Designation"},"date_of_joining":{"type":"string","format":"date","title":"Date Of Joining"},"phone":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Phone"},"is_active":{"type":"boolean","title":"Is Active"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"updated_at":{"type":"string","format":"date-time","title":"Updated At"}},"type":"object","required":["id","employee_code","name","email","department","designation","date_of_joining","phone","is_active","created_at","updated_at"],"title":"EmployeeResponse","description":"Response schema for employee data."},"EmployeeUpdate":{"properties":{"name":{"anyOf":[{"type":"string","maxLength":100,"minLength":1},{"type":"null"}],"title":"Name"},"email":{"anyOf":[{"type":"string","maxLength":255,"minLength":5},{"type":"null"}],"title":"Email"},"department":{"anyOf":[{"type":"string","maxLength":100,"minLength":1},{"type":"null"}],"title":"Department"},"designation":{"anyOf":[{"type":"string","maxLength":100},{"type":"null"}],"title":"Designation"},"date_of_joining":{"anyOf":[{"type":"string","format":"date"},{"type":"null"}],"title":"Date Of Joining"},"phone":{"anyOf":[{"type":"string","maxLength":20},{"type":"null"}],"title":"Phone"},"is_active":{"anyOf":[{"type":"boolean"},{"type":"null"}],"title":"Is Active"}},"type":"object","title":"EmployeeUpdate","description":"Request schema for updating an employee. All fields optional."},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"PaginatedResponse_AttendanceResponse_":{"properties":{"data":{"items":{"$ref":"#/components/schemas/AttendanceResponse"},"type":"array","title":"Data"},"meta":{"$ref":"#/components/schemas/PaginationMeta"}},"type":"object","required":["data","meta"],"title":"PaginatedResponse[AttendanceResponse]"},"PaginatedResponse_EmployeeResponse_":{"properties":{"data":{"items":{"$ref":"#/components/schemas/EmployeeResponse"},"type":"array","title":"Data"},"meta":{"$ref":"#/components/schemas/PaginationMeta"}},"type":"object","required":["data","meta"],"title":"PaginatedResponse[EmployeeResponse]"},"PaginationMeta":{"properties":{"page":{"type":"integer","title":"Page"},"per_page":{"type":"integer","title":"Per Page"},"total":{"type":"integer","title":"Total"},"total_pages":{"type":"integer","title":"Total Pages"}},"type":"object","required":["page","per_page","total","total_pages"],"title":"PaginationMeta","description":"Pagination metadata for list endpoints."},"StatusSummary":{"properties":{"present":{"type":"integer","title":"Present","default":0},"absent":{"type":"integer","title":"Absent","default":0},"half_day":{"type":"integer","title":"Half Day","default":0},"on_leave":{"type":"integer","title":"On Leave","default":0}},"type":"object","title":"StatusSummary","description":"Aggregated counts per attendance status."},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}